+++ server/tests-hspec/Harness/Backend/BigQuery.hs
@@ -6,9 +6,11 @@
 -- | BigQuery helpers.
 module Harness.Backend.BigQuery
   ( run_,
+    runSql_,
     getServiceAccount,
     getProjectId,
     createTable,
+    defaultSourceMetadata,
     insertTable,
     trackTable,
     dropTable,
@@ -18,6 +20,8 @@ module Harness.Backend.BigQuery
   )
 where
 
+import Control.Monad (void)
+import Data.Aeson (Value)
 import Data.Bool (bool)
 import Data.Foldable (for_)
 import Data.String
@@ -28,13 +32,13 @@ import Data.Time (defaultTimeLocale, formatTime)
 import GHC.Stack
 import Harness.Constants as Constants
 import Harness.Env
-import Harness.Exceptions (forFinally_)
+import Harness.Exceptions
 import Harness.GraphqlEngine qualified as GraphqlEngine
 import Harness.Quoter.Yaml (yaml)
-import Harness.State (State)
 import Harness.Test.Context (BackendType (BigQuery), defaultBackendTypeString, defaultSource)
-import Harness.Test.Schema (BackendScalarType (..), BackendScalarValue (..), ScalarValue (..), formatBackendScalarValue)
+import Harness.Test.Schema (BackendScalarType (..), BackendScalarValue (..), ScalarValue (..))
 import Harness.Test.Schema qualified as Schema
+import Harness.TestEnvironment (TestEnvironment)
 import Hasura.Backends.BigQuery.Connection (initConnection)
 import Hasura.Backends.BigQuery.Execute qualified as Execute
 import Hasura.Backends.BigQuery.Source (ServiceAccount)
@@ -54,6 +58,31 @@ run_ serviceAccount projectId query = do
   conn <- initConnection serviceAccount projectId Nothing
   res <- Execute.executeBigQuery conn Execute.BigQuery {Execute.query = fromString query, Execute.parameters = mempty}
   res `onLeft` (`bigQueryError` query)
+
+runSql_ :: HasCallStack => String -> IO ()
+runSql_ query = do
+  serviceAccount <- getServiceAccount
+  projectId <- getProjectId
+  catch
+    ( bracket
+        (initConnection serviceAccount projectId Nothing)
+        (const (pure ()))
+        (\conn -> void $ handleResult <$> (Execute.executeBigQuery conn Execute.BigQuery {Execute.query = fromString query, Execute.parameters = mempty}))
+    )
+    ( \(e :: SomeException) ->
+        error
+          ( unlines
+              [ "BigQuery error:",
+                show e,
+                "SQL was:",
+                query
+              ]
+          )
+    )
+  where
+    handleResult :: Either Execute.ExecuteProblem () -> IO ()
+    handleResult (Left _) = throwString "Error handling bigquery"
+    handleResult (Right ()) = pure ()
 
 bigQueryError :: HasCallStack => Execute.ExecuteProblem -> String -> IO ()
 bigQueryError e query =
@@ -133,7 +162,7 @@ serialize = \case
   VUTCTime t -> pack $ formatTime defaultTimeLocale "'%F %T'" t
   VBool b -> tshow @Int $ if b then 1 else 0
   VNull -> "NULL"
-  VCustomValue bsv -> "'" <> formatBackendScalarValue bsv bsvBigQuery <> "'"
+  VCustomValue bsv -> Schema.formatBackendScalarValueType $ Schema.backendScalarValue bsv bsvBigQuery
 
 mkRow :: [Schema.ScalarValue] -> Text
 mkRow row =
@@ -160,12 +189,12 @@ dropTable Schema.Table {tableName} = do
 
 -- | Post an http request to start tracking
 -- Overriding here because bigquery's API is uncommon
-trackTable :: State -> Schema.Table -> IO ()
-trackTable state Schema.Table {tableName} = do
+trackTable :: TestEnvironment -> Schema.Table -> IO ()
+trackTable testEnvironment Schema.Table {tableName} = do
   let datasetName = T.pack Constants.bigqueryDataset
       source = defaultSource BigQuery
   GraphqlEngine.postMetadata_
-    state
+    testEnvironment
     [yaml|
 type: bigquery_track_table
 args:
@@ -177,12 +206,12 @@ args:
 
 -- | Post an http request to stop tracking the table
 -- Overriding `Schema.trackTable` here because bigquery's API expects a `dataset` key
-untrackTable :: State -> Schema.Table -> IO ()
-untrackTable state Schema.Table {tableName} = do
+untrackTable :: TestEnvironment -> Schema.Table -> IO ()
+untrackTable testEnvironment Schema.Table {tableName} = do
   let datasetName = T.pack Constants.bigqueryDataset
       source = defaultSource BigQuery
   GraphqlEngine.postMetadata_
-    state
+    testEnvironment
     [yaml|
 type: bigquery_untrack_table
 args:
@@ -192,10 +221,33 @@ args:
     name: *tableName
 |]
 
+-- | Metadata source information for the default BigQuery instance
+defaultSourceMetadata :: IO Value
+defaultSourceMetadata = do
+  let dataset = Constants.bigqueryDataset
+      source = defaultSource BigQuery
+      backendType = defaultBackendTypeString BigQuery
+  serviceAccount <- getServiceAccount
+  projectId <- getProjectId
+  pure $
+    [yaml|
+type: replace_metadata
+args:
+  version: 3
+  sources:
+  - name: *source
+    kind: *backendType
+    tables: []
+    configuration:
+      service_account: *serviceAccount
+      project_id: *projectId
+      datasets: [*dataset]
+|]
+
 -- | Setup the schema in the most expected way.
 -- NOTE: Certain test modules may warrant having their own local version.
-setup :: [Schema.Table] -> (State, ()) -> IO ()
-setup tables (state, _) = do
+setup :: [Schema.Table] -> (TestEnvironment, ()) -> IO ()
+setup tables (testEnvironment, _) = do
   let dataset = Constants.bigqueryDataset
       source = defaultSource BigQuery
       backendType = defaultBackendTypeString BigQuery
@@ -203,7 +255,7 @@ setup tables (state, _) = do
   serviceAccount <- getServiceAccount
   projectId <- getProjectId
   GraphqlEngine.postMetadata_
-    state
+    testEnvironment
     [yaml|
 type: replace_metadata
 args:
@@ -221,20 +273,20 @@ args:
   for_ tables $ \table -> do
     createTable table
     insertTable table
-    trackTable state table
+    trackTable testEnvironment table
   -- Setup relationships
   for_ tables $ \table -> do
-    Schema.trackObjectRelationships BigQuery table state
-    Schema.trackArrayRelationships BigQuery table state
+    Schema.trackObjectRelationships BigQuery table testEnvironment
+    Schema.trackArrayRelationships BigQuery table testEnvironment
 
 -- | Teardown the schema and tracking in the most expected way.
 -- NOTE: Certain test modules may warrant having their own version.
-teardown :: [Schema.Table] -> (State, ()) -> IO ()
-teardown (reverse -> tables) (state, _) = do
+teardown :: [Schema.Table] -> (TestEnvironment, ()) -> IO ()
+teardown (reverse -> tables) (testEnvironment, _) = do
   -- Teardown relationships first
   forFinally_ tables $ \table ->
-    Schema.untrackRelationships BigQuery table state
+    Schema.untrackRelationships BigQuery table testEnvironment
   -- Then teardown tables
   forFinally_ tables $ \table -> do
-    untrackTable state table
+    untrackTable testEnvironment table
     dropTable table